Conflicts:
	App/EzService/EzService/ActionResults/NewLoanModelActionResult.cs
	App/EzServiceClient/Service References/EzServiceReference/Ezbob.Backend.Models.NewLoan.xsd
	App/EzServiceClient/Service References/EzServiceReference/Ezbob.Backend.Models.xsd
	App/EzServiceClient/Service References/EzServiceReference/Ezbob.Backend.ModelsWithDB.NewLoan.xsd
	App/EzServiceClient/Service References/EzServiceReference/Reference.cs
	App/EzServiceClient/Service References/EzServiceReference/Reference.svcmap
	App/EzServiceClient/Service References/EzServiceReference/configuration.svcinfo
	App/EzServiceClient/Service References/EzServiceReference/configuration91.svcinfo
	App/EzServiceClient/Service References/EzServiceReference/service.wsdl
	App/EzServiceClient/Service References/EzServiceReference/service.xsd
	App/EzServiceClient/app.config
	App/PluginWeb/EzBob.Web/Areas/Underwriter/Controllers/CustomersReview/ApplicationInfoController.cs
	App/PluginWeb/EzBob.Web/Areas/Underwriter/Controllers/LoanEditorController.cs
	App/PluginWeb/EzBob.Web/Code/CashRequestBuilder.cs
	App/PluginWeb/EzBob.Web/EzBob.Web.csproj
	Backend/CalculateLoan/LoanCalculator/ALoanCalculator.cs
	Backend/Strategies/MainStrategy/MainStrategy.cs
	Backend/Strategies/NewLoan/RescheduleLoan.cs
	Integration/EzBob/EzBobTest/Misc.cs
	Integration/EzBob/EzBobTest/TestStrategies.cs
	Integration/PaymentServices/Calculators/APRCalculator.cs
	sql/Current/2015_04_28_01_CreditSafeNonLtdComments.sql
	sql/Current/2015_05_12_01_NewLoan.sql
	sql/Current/2015_07_27_01_NewLoan.sql
	sql/old_current/2015-06-01/2015_04_28_01_CreditSafeNonLtdComments.sql
	sql/old_current/2015-07-19/2015_04_28_01_CreditSafeNonLtdComments.sql
	sql/old_current/2015-07-19/2015_05_12_01_NewLoan.sql
	sql/old_current/2015-07-19/2015_05_20_01_NL_LoanAgreements_AgreementModel.sql

+++ Tools/EzbobUtils/AutomationVerification/AutomationVerification/AutoDecision/AutoReApproval/AutoReApprovalModels.cs
@@ -66,7 +66,7 @@ namespace AutomationCalculator.AutoDecision.AutoReApproval
 		public bool HasLoanCharges { get; set; }
 		public decimal ReApproveAmount { get; set; }
 		public decimal AvaliableFunds { get; set; }
-        public int LacrID { get; set; }
+        public long LacrID { get; set; }
 
 		//configs
 		public int AutoReApproveMaxLacrAge { get; set; }
